Moved examples_helper into openff/toolkit, fixed LGTM
