+++ openff/toolkit/examples_helper/example.py
@@ -1,9 +1,10 @@
 from pathlib import Path
 import click
 
-OFFTK_ROOT = (Path(__file__) / "../../../").resolve()
+OFFTK_ROOT = (Path(__file__) / "../../../../").resolve()
 EXAMPLES_DIR = OFFTK_ROOT / "examples"
 EXAMPLES_ENV = EXAMPLES_DIR / "environment.yaml"
+
 
 class Example:
     """
